refac: Cleanup application folder structure
