+++ source/dotnet/Services/Application/Batches/BatchExecutionStateHandler.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.using Energinet.DataHub.Wholesale.Application.JobRunner;
 
-using Energinet.DataHub.Wholesale.Application.JobRunner;
+using Energinet.DataHub.Wholesale.Application.CalculationJobs;
 using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 using Microsoft.Extensions.Logging;
 using NodaTime;
