+++ source/dotnet/Services/Application/Batches/IBatchCompletedPublisher.cs
