+++ source/dotnet/Services/Application/CalculationJobs/ICalculatorJobRunner.cs
@@ -12,13 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Application.JobRunner;
+using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 
-public enum JobState
+namespace Energinet.DataHub.Wholesale.Application.CalculationJobs;
+
+public interface ICalculatorJobRunner
 {
-    Pending,
-    Running,
-    Completed,
-    Canceled,
-    Failed,
+    Task<JobState> GetJobStateAsync(JobRunId jobRunId);
+
+    /// <summary>
+    /// Start job.
+    /// </summary>
+    /// <param name="jobParameters">
+    /// Parameters must be on the form "--param-name=param-value".
+    /// Further details about the format depends on the actual job runner implementation.
+    /// </param>
+    Task<JobRunId> SubmitJobAsync(IEnumerable<string> jobParameters);
 }
