+++ source/dotnet/Services/Application/Processes/IProcessCompletedIntegrationEventPublisher.cs
