+++ source/dotnet/Services/Infrastructure/Processes/ProcessResultPoint.cs
